move helper to utils module
